+++ src/fzf/azext_fzf/__init__.py
@@ -2,6 +2,9 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
+"""
+Initializes the azext_fzf module.
+"""
 
 from azure.cli.core import AzCommandsLoader
 
@@ -9,15 +12,16 @@ from azext_fzf._help import helps  # pylint: disable=unused-import
 
 
 class FzfCommandsLoader(AzCommandsLoader):
+    """
+    Class to load the commands and arguments for the fzf extension.
+    """
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
-        from azext_fzf._client_factory import cf_fzf
         fzf_custom = CliCommandType(
-            operations_tmpl='azext_fzf.custom#{}',
-            client_factory=cf_fzf)
+            operations_tmpl='azext_fzf.custom#{}')
         super(FzfCommandsLoader, self).__init__(cli_ctx=cli_ctx,
-                                                  custom_command_type=fzf_custom)
+                                                custom_command_type=fzf_custom)
 
     def load_command_table(self, args):
         from azext_fzf.commands import load_command_table
@@ -29,4 +33,4 @@ class FzfCommandsLoader(AzCommandsLoader):
         load_arguments(self, command)
 
 
-COMMAND_LOADER_CLS = FzfCommandsLoader
+COMMAND_LOADER_CLS = FzfCommandsLoader # pylint: disable=invalid-name