+++ src/fzf/azext_fzf/_help.py
@@ -3,6 +3,9 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
+"""
+Extension help for azext_fzf.
+"""
 
 from knack.help_files import helps  # pylint: disable=unused-import