+++ src/fzf/HISTORY.rst
@@ -10,3 +10,9 @@ Release History
 0.1.1
 ++++++
 * Added better help, fixed dependency info, added better error handling.
+
+0.2.0
+++++++
+* Implemented fzf functionality internally instead of using a module
+* Initial unit testing
+* Implemented install functionality